Removed TrieUtil.java and moved to BaseCommand
